various edits, and removed EntityComponent jargon
More specifically, I deleted EntityComponentProvider and other things and replaced it with just EntityHook, and left HookLoader. The reason being that 1) it's a little overcomplicated, but 2) if we really need it, mirsar's working on conditional instancing for 1.4. Performance wise, it's not a major concern anyway, and the new system will be far better than anything I could have come up with.